+++ src/features/dashboard/components/asidebar/Asidebar.scss
@@ -1,0 +1,16 @@
+.asidebar {
+  &__nav {
+    flex-basis: 212px;
+    box-shadow: none;
+
+    &-item {
+      &_active {
+        color: var(--button-active-color);
+      }
+
+      &_type_exit {
+        color: var(--button-super-red);
+      }
+    }
+  }
+}
